+++ include/bedrock/world/math/vec3.h
@@ -19,5 +19,25 @@ public:
     float x;
     float y;
     float z;
+
+    constexpr Vec3 operator+(const Vec3 &other) const
+    {
+        return {x + other.x, y + other.y, z + other.z};
+    }
+
+    constexpr Vec3 operator-(const Vec3 &other) const
+    {
+        return {x - other.x, y - other.y, z - other.z};
+    }
+
+    constexpr Vec3 operator*(const Vec3 &other) const
+    {
+        return {x * other.x, y * other.y, z * other.z};
+    }
+
+    constexpr Vec3 operator/(const Vec3 &other) const
+    {
+        return {x / other.x, y / other.y, z / other.z};
+    }
 };
 static_assert(sizeof(Vec3) == 12);