+++ src/endstone_runtime/bedrock/world/actor/actor_collision.cpp
@@ -12,20 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
+#include "bedrock/world/actor/actor_collision.h"
 
-template <typename ComponentFlag>
-class FlagComponent {
-public:
-    ComponentFlag flag;
-};
+#include "bedrock/world/actor/components/flag_component.h"
 
-struct PlayerComponentFlag {};
-struct OnGroundFlag {};
-struct WasOnGroundFlag {};
+namespace ActorCollision {
 
-#if defined __clang__ || defined __GNUC__
-static_assert(entt::type_hash<FlagComponent<PlayerComponentFlag>>::value() == 0x356FFFCA);
-#elif defined _MSC_VER
-static_assert(entt::type_hash<FlagComponent<PlayerComponentFlag>>::value() == 0x600FBE1D);
-#endif
+bool isOnGround(const EntityContext &ctx)
+{
+    return ctx.hasComponent<FlagComponent<OnGroundFlag>>();
+}
+
+bool wasOnGround(const EntityContext &ctx)
+{
+    return ctx.hasComponent<FlagComponent<WasOnGroundFlag>>();
+}
+
+}  // namespace ActorCollision