+++ CHANGELOG.md
@@ -14,12 +14,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Server::getOnlinePlayers` for getting all online players.
 - `Player::kick` for kicking a specific player with a custom message.
 - `Player::getPing` for getting the average latency of a connected player.
+- `Actor::getLocation` and `Actor::getVelocity` for getting the location and velocity of an actor respectively.
+- `Actor::getLevel` for getting the `Level` that an actor resides in.
 
 ### Changed
 
 - `Level` is no longer created for each `Dimension` in a Minecraft world. Instead, one `Level` now manages multiple
   `Dimension`s.
-- `Player::getAddress` now returns a `SocketAddress` instead of `std::string`. `Player::getPort` was removed, use `SocketAddress::getPort` instead.
+- `Player::getAddress` now returns a `SocketAddress` instead of `std::string`. `Player::getPort` was removed,
+  use `SocketAddress::getPort` instead.
 
 ## [0.4.2](https://github.com/EndstoneMC/endstone/releases/tag/v0.4.2) - 2024-04-30